Provide default ip address and user agent on error
Previously a panic would occur in the following two cases:

Requests with no user agent are valid according to the HTTP spec RFC 2616 and 7231. In such a case, the default "No user agent" will be used instead.

IP address parsing would fail if the connection came from `localhost` or a similar address. In such a case, an IPv6 black hole IP (`100::/64`) will be used as the default instead.