Implement `ErrorHolder` wrapping struct and `HoldsError` trait
This is for usage with generic methods which have an extra `Error` type. Defining that to the `anyhow::Error` results in generic methods being unable to restrict the `Error` to an instance of a `anyhow::Error` and thus being unable to use it.

By defining the `Error` type as `ErrorHolder` (which is a simple `#[repr(transparent)]` struct which just wraps an `anyhow::Error`), the trait bound for the `Error` type can be defined as `HoldsError`. This trait is guaranteed to only be implemented for `ErrorHolder`. One can then receive the original error by calling `.into_inner()` (consuming) or `.as_inner()` (referencing).